trestle-cli/pyproject.toml
# Added to aide with eventual transition away from setup.py
[build-system]
requires = [
    "setuptools >= 35.0.2",
    "wheel",
    "setuptools_scm >= 2.0.0"
]
build-backend = "setuptools.build_meta"

[project]
name = "compliance-trestle"
description = "Tools to manage & autogenerate python objects representing the OSCAL layers/models"
authors = [
    {name = "OSCAL Compass", email = "oscal-compass-oversight@googlegroups.com"}
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "Operating System :: POSIX",
    "Operating System :: Microsoft",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
readme = "README.md"
license = "Apache-2.0"
license-files = ["LICENSE"]
dynamic = ["version"]
requires-python = ">=3.9"
dependencies = [
    "attrs",
    "ilcli",
    "cryptography==44.0.2",
    "paramiko==3.5.0",
    "ruamel.yaml",
    "furl",
    "pydantic[email]>=2.0.0",
    "python-dotenv>=0.10.4",
    "datamodel-code-generator[http] == 0.25.3",
    "python-frontmatter",
    "pywin32 >= 1.0;platform_system=='Windows'",
    "defusedxml",
    "openpyxl~=3.0",
    "Jinja2 == 3.1.6",
    "cmarkgfm>=2024.1,<2024.12", ## Update regularly
    "orjson",
    "requests>=2.32.2",
    "importlib_resources",
    "trestle-sdk",
]

[project.urls]
Homepage = "https://oscal-compass.github.io/compliance-trestle"

[project.scripts]
trestle = "trestle.cli:run"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["trestle*"]
exclude = ["tests"]

[tool.setuptools.package-data]
trestle = ["*.ini", "*.md", "*.jinja", "*.drawio", "*.json", "*.yaml", "*.yml"]

[tool.setuptools.dynamic]
version = {attr = "trestle.__init__.__version__"}

[tool.pytest.ini_options]
minversion = "6.2"
testpaths = [
    "tests",
]

[tool.black]
line-length = 500

[tool.isort]
line_length = 500

[tool.flake8]
# WARNING: This should be kept compatible with .pre-commit-config.yaml
ignore = ["P1", "C812", "C813", "C814", "C815", "C816"]
max-line-length = 120

[tool.mypy]
plugins = "pydantic.v1.mypy"
ignore_missing_imports = true
strict_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
disallow_any_generics = true
check_untyped_defs = true
no_implicit_reexport = true
show_error_codes = true
show_error_context = true
disallow_untyped_defs = true
disable_error_code = ["union-attr", "attr-defined", "no-redef", "assignment", "arg-type", "list-item"]


[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true
warn_untyped_fields = true

[tool.coverage.run]
relative_files = true

[tool.semantic_release]
build_command = """
    python -m pip install -e .[dev] --upgrade --upgrade-strategy eager --
    python setup.py sdist bdist_wheel
"""
version_variables = ["trestle/__init__.py:__version__"]
commit_author = "semantic-release <semantic-release>"

[tool.semantic_release.commit_parser_options]
allowed_tags = ["build", "chore", "ci", "docs", "feat", "fix", "perf", "style", "refactor", "test"]
minor_tags = ["feat"]
patch_tags = ["fix", "perf"]
default_bump_level = 0

[tool.semantic_release.branches.main]
match = "(main)"
prerelease_token = "rc"
prerelease = false

[tool.semantic_release.remote]
name = "origin"
type = "github"
ignore_token_for_push = false
insecure = false

[tool.semantic_release.publish]
dist_glob_patterns = ["dist/*"]

[tool.uv.sources]
trestle-sdk = { workspace = true }
